refactor: Rename 'Viewer' to 'View'
