+++ frontend/src/components/View/CanvasContainer.js
