+++ frontend/src/components/View/DualView.js
