+++ frontend/src/components/View/DualViewHeader.js
