+++ frontend/src/components/View/SingleView.js
