+++ frontend/src/components/View/ViewTitle.js
