+++ frontend/src/components/View/WelcomeMessage.js
